Add files via upload
This code is for an ESP8266-based irrigation controller that can be accessed through a web interface. The controller can be set up with a watering schedule, and it can check the local weather forecast to determine if it should water the plants or not. The code also includes features such as OTA updates and an I2C 16x2 LCD screen to display the time and system information.